add top-level wrappers for 16b sprintz

+++ cpp/Compress/sprintz_xff_rle.cpp
@@ -1175,8 +1175,8 @@ SPRINTZ_FORCE_INLINE int64_t decompress_rowmajor_xff_rle_8b(const int8_t* src,
 {
     return decompress_rowmajor_xff_rle(src, dest, ndims, ngroups, remaining_len);
 }
-SPRINTZ_FORCE_INLINE int64_t decompress_rowmajor_xff_rle_16b(const int8_t* src,
-    uint8_t* dest, uint16_t ndims, uint32_t ngroups, uint16_t remaining_len)
+SPRINTZ_FORCE_INLINE int64_t decompress_rowmajor_xff_rle_16b(const int16_t* src,
+    uint16_t* dest, uint16_t ndims, uint32_t ngroups, uint16_t remaining_len)
 {
     return decompress_rowmajor_xff_rle(src, dest, ndims, ngroups, remaining_len);
 }
